(feature)[boot/fat12]: Add a basic fat12 read(WIP)
* Taken the basic shape of the fat12 parser my repo: git@github.com:TrazelDev/fat12-parser.git and moved it the bootloader.
* After moving made sure all of the libc functions that the parser used are implemented (malloc, tolower)

What next:
* Implement better abstraction for ATA driver.
* Need to implement a function that reads from the hard drive that is a wrapper function around my ATA driver

+++ boot/bootloader/bootloader.c
@@ -20,7 +20,7 @@ void bootloader_entry() {
 	allocator.init((void*)0x20000, (void*)0x20500);
 	char* addr = allocator.malloc(0x100);
 	addr = allocator.malloc(200);
-	itoa_unsigned(addr, addr, HEX);
+	itoa_unsigned((uint64_t)addr, addr, HEX);
 	print_string("allocated addr: 0x");
 	print_string(addr);
 	print_string("\n");
